+++ src/engine/Sprite.ts
